Latest version

+++ CoderGirl_MVCMovies/Data/IMovieRatingRepository.cs
@@ -5,9 +5,8 @@ using System.Threading.Tasks;
 
 namespace CoderGirl_MVCMovies.Data
 {
-    // TODO: Implement this interface
     public interface IMovieRatingRepository
-    {        
+    {
         /// <summary>
         /// Given a movieName and rating, saves the rating and returns a unique id > 0.
         /// If the movie name and/or rating are null or empty, nothing should be saved and it should return 0
@@ -19,7 +18,7 @@ namespace CoderGirl_MVCMovies.Data
 
         /// <summary>
         /// Given an id, will return the associated rating 
-        /// If the id does not exist, returns null
+        /// If the id does not exist, returns 0
         /// </summary>
         /// <param name=""></param>
         /// <returns></returns>
@@ -39,7 +38,7 @@ namespace CoderGirl_MVCMovies.Data
         /// </summary>
         /// <param name="movieName"></param>
         /// <returns></returns>
-        decimal GetAverageRatingByMovieName(string movieName);
+        double GetAverageRatingByMovieName(string movieName);
 
         /// <summary>
         /// Returns a list of all the ids of saved movie ratings
